# apps/castaway-web/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Castaway Web API
  version: 0.0.0
...
paths:
...
  /instances/{instanceID}/drafts/{participantID}/late:
    post:
      operationId: acceptLateDraft
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
        - name: participantID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    additionalProperties: {}
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceDraftRequest'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    ReplaceDraftRequest:
      type: object
      required:
        - contestant_ids
      properties:
        contestant_ids:
          type: array
          items:
            type: string
        idempotency_key:
          type: string
        effective_at:
          type: string
          format: date-time
        reason:
          type: string
...
servers:
  - url: http://localhost:8080
    description: Local development
    variables: {}